Created and returned list of threads via API
