+++ frontend/frontend/lib/src/screens/noticeboardScreen.dart
@@ -1,7 +1,3 @@
-import 'dart:convert';
-import 'dart:developer';
-import 'dart:io';
-import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 import 'package:frontend/src/models/util_model.dart';
 import 'package:frontend/src/screens/userProfileScreen.dart';
@@ -15,94 +11,12 @@ class NoticeBoard extends StatefulWidget {
   }
 }
 
-class NoticeBoardThreads {
-  final List<dynamic> threadList;
-
-  NoticeBoardThreads({
-    required this.threadList,
-  });
-
-  factory NoticeBoardThreads.fromJson(Map<String, dynamic> json) {
-    return NoticeBoardThreads(
-         threadList: json["noticeBoard"],
-    );
-  }
-}
-
-class Thread {
-  final int threadId;
-  final String threadTitle;
-  final String threadContent;
-  final int minLevel;
-  final String imageUrl;
-  final int permittedUserRoles;
-  final int userId;
-
-  Thread({
-    required this.threadId,
-    required this.threadTitle,
-    required this.threadContent,
-    required this.minLevel,
-    required this.imageUrl,
-    required this.permittedUserRoles,
-    required this.userId,
-  });
-
-  factory Thread.fromJson(Map<String, dynamic> json) {
-    return Thread(
-      threadId: json['threadId'],
-      threadTitle: json['threadTitle'],
-      threadContent: json['threadContent'],
-      minLevel: json['minLevel'],
-      imageUrl: json['imageUrl'],
-      permittedUserRoles: json['permittedUserRoles'],
-      userId: json['userId'],
-    );
-  }
-}
-
-Future<NoticeBoardThreads> fetchNotice() async {
-
-    HttpClient client = new HttpClient();
-    client.badCertificateCallback = ((X509Certificate cert, String host, int port) => true);
-    String url ='http://10.0.2.2:5000/api/NoticeBoard/RetrieveNoticeBoardThreads';
-    //Map map = { "email" : "email" , "password" : "password" };
-    HttpClientRequest request = await client.getUrl(Uri.parse(url));
-    request.headers.set('content-type', 'application/json');
-    //request.add(utf8.encode(json.encode(map)));
-    HttpClientResponse response1 = await request.close();
-    String reply = await response1.transform(utf8.decoder).join();
-
-    print(NoticeBoardThreads.fromJson(jsonDecode(reply)).threadList);
-    print(NoticeBoardThreads.fromJson(jsonDecode(reply)).threadList[0]);
-
-
-    var test = (NoticeBoardThreads.fromJson(jsonDecode(reply)).threadList[0]);
-    print("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
-    print(test);
-    print(jsonEncode(test));
-
-    print(Thread.fromJson(test).threadId);
-
-
-
-    return NoticeBoardThreads.fromJson(jsonDecode(reply));
-
-}
-
 class _NoticeBoard extends State<NoticeBoard> {
   final util = new UtilModel();
-  late Future<NoticeBoardThreads> futureNotice;
 
-  @override
-  void initState() {
-    super.initState();
-    futureNotice = fetchNotice();
-  }
   void next() {
     UtilModel.route(() => ProfileScreen(), context);
   }
-
 
   @override
   Widget build(BuildContext context) {
@@ -126,9 +40,6 @@ class _NoticeBoard extends State<NoticeBoard> {
       ),
       floatingActionButton: ExampleExpandableFab(),
       backgroundColor: Color.fromRGBO(33, 33, 33, 1),
-
-
-
       body:
       Center(
         child: Stack(
@@ -140,18 +51,6 @@ class _NoticeBoard extends State<NoticeBoard> {
              Container(
                 child: NoticeboardCard(),
               ),
-            FutureBuilder<NoticeBoardThreads>(
-              future: futureNotice,
-              builder: (context, snapshot) {
-                if (snapshot.hasData) {
-                  return Text(snapshot.data!.toString());
-                } else if (snapshot.hasError) {
-                  return Text("${snapshot.error}");
-                }
-                // By default, show a loading spinner.
-                return Text(snapshot.data.toString());
-              },
-            )
           ],
         ),
       ),
